WIP: move propose.go into core.go
